Struct cint::Alpha

#[repr(C)]
pub struct Alpha<ColorTy: ColorType> {
    pub color: ColorTy,
    pub alpha: ColorTy::ComponentTy,
}

A color with an alpha component.

The color components and alpha component are completely separate.

Fields
color: ColorTy

The contained color, which is completely separate from the alpha value.

alpha: ColorTy::ComponentTy

The alpha component.