Struct cint::PremultipliedAlpha

pub struct PremultipliedAlpha<ColorTy: ColorType> {
    pub color: ColorTy,
    pub alpha: ColorTy::ComponentTy,
}

A premultiplied color with an alpha component.

The color components have been premultiplied by the alpha component.

Fields
color: ColorTy

The contained color, which has been premultiplied with alpha

alpha: ColorTy::ComponentTy

The alpha component.